[processor] cmc api consumer and discord supplier for fear and greed indicator
